+++ docs/examples/actions/quiter.py
@@ -2,7 +2,7 @@ from textual.app import App
 
 
 class Quiter(App):
-    async def on_load(self, event):
+    async def on_load(self):
         await self.bind("q", "quit")