+++ docs/examples/messages_and_events/color_changer.py
@@ -2,7 +2,7 @@ from textual.app import App
 
 
 class ColorChanger(App):
-    async def on_key(self, event):
+    def on_key(self, event):
         if event.key.isdigit():
             self.background = f"on color({event.key})"