+++ examples/grid.py
@@ -1,10 +1,9 @@
 from textual.app import App
-from textual import events
 from textual.widgets import Placeholder
 
 
 class GridTest(App):
-    async def on_mount(self, event: events.Mount) -> None:
+    async def on_mount(self) -> None:
         """Make a simple grid arrangement."""
 
         grid = await self.view.dock_grid(edge="left", size=70, name="left")