+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1.9] - Unreleased
+## [0.1.10] - Unreleased
+
+### Changed
+
+- Callbacks may be async or non-async.
+- Event handler event argument is optional.
+- Fixed exception in clock example https://github.com/willmcgugan/textual/issues/52
+- Added Message.wait() which waits for a message to be processed
+
+## [0.1.9] - 2021-08-06
 
 ### Added